Added duplicate block comments to ensure their output—TEMPORARY WORKAROUND!
Dupes are evil.  But it’s important to me that the comments are output, and Stylus is suppressing comments in `@import`ed files if those comments are not inside a selector.  It’s even suppressing `/*!` prefixed block comments, which I’m pretty sure is a bug.  I opened an issue here:

https://github.com/LearnBoost/stylus/issues/444